Fix potential space leak when generating in-order results
This confounded me for WEEKS. I was almost convinced I'd come across
a GHC 7.6 bug, as this manifested as a stack overflow on my office
machine, which I was unable to reproduce on my (GHC 7.8 using) desktop.

Only when trying to do some benchmarks did I notice the huge performance
disparity between 3 and 4 threads on the AmicablePairs benchmark on my
home machine, and so I brought out the profiling tools and noticed a
*huge* space leak of AP_STACK nodes in the 4-thread run. As it only
appeared when generating in-order streams, I made a lucky guess, and
fixed it!

Huzzah!

Coincidentally, I had just been reading Small Gods, so I am willing to
consider this a miracle given to me by an unknown local deity.