Improve worst-case performance.
The worst-case with the old implementation is that there are n work
items and n parallel agents. The first agent would claim all work,
and the others would have no work to do. The approach of claiming
n work items works well if the amount of work is large compared to
the number of agents, but fails when there are few.

Rather than claim n, we can claim 1 on the first iteration, 2 on
the second, 3 on the third, and so on, up to n.